Solve in-process symbol problem
Create a python script that emulates being julia and simply hands off control.
This is necessary because the precompilation and runtime environment needs to be
the same in order for PyCall to be able to be used without major modification.

Try a different approach

Only use hack if python is statically linked

New version yet again