Set --startup-file=no in juliainfo
Currently, startup file is not loaded in the main PyJulia session.
Including the startup file has some potential problem.  For example,
if something is printed in the startup file then `juliainfo` would
break.  So it seems to be that the safe default is to avoid loading
the startup file (though it makes sense to add an option to load it in
the future).