fix `sysimage` argument to `julia.core.Julia.__init__`
On `master` branch, it should be possible to pass a custom "sysimage" to
the `julia.core.Julia.__init__` method. However, if the python version
in use is statically linked, an UnsupportedPythonError is thrown
(despite the sysimage argument having been passed to `__init__`).

This commit prevents the UnsupportedPythonError from being thrown if a
sysimage keyword argument is passed to `__init__`.
Closes #421